Use native crypto only if it supports DH.
