+++ lib/crypto-browser.js
@@ -1,11 +1,4 @@
-import nacl from 'tweetnacl';
-
-export async function deriveSecret({privateKey, remotePublicKey}) {
-  return nacl.scalarMult(privateKey, remotePublicKey);
-}
-
-export async function generateKeyPair() {
-  // Each is a Uint8Array with 32-byte key
-  const {publicKey, secretKey: privateKey} = nacl.box.keyPair();
-  return {publicKey, privateKey};
-}
+/*!
+ * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ */
+export {deriveSecret, generateKeyPair} from './crypto-nacl.js';
