+++ lib/crypto.js
@@ -3,6 +3,7 @@
  */
 import * as crypto from 'crypto';
 import {promisify} from 'util';
+import * as cryptoNacl from './crypto-nacl.js';
 
 const PUBLIC_KEY_DER_PREFIX = new Uint8Array([
   48, 42, 48, 5, 6, 3, 43, 101, 110, 3, 33, 0
@@ -12,7 +13,7 @@ const PRIVATE_KEY_DER_PREFIX = new Uint8Array([
   48, 46, 2, 1, 0, 48, 5, 6, 3, 43, 101, 110, 4, 34, 4, 32
 ]);
 
-export async function deriveSecret({privateKey, remotePublicKey}) {
+async function deriveSecretNative({privateKey, remotePublicKey}) {
   const nodePrivateKey = crypto.createPrivateKey({
     key: Buffer.concat([PRIVATE_KEY_DER_PREFIX, privateKey]),
     format: 'der',
@@ -29,7 +30,7 @@ export async function deriveSecret({privateKey, remotePublicKey}) {
   });
 }
 
-export async function generateKeyPair() {
+async function generateKeyPairNative() {
   const generateKeyPairAsync = promisify(crypto.generateKeyPair);
   const publicKeyEncoding = {format: 'der', type: 'spki'};
   const privateKeyEncoding = {format: 'der', type: 'pkcs8'};
@@ -41,3 +42,14 @@ export async function generateKeyPair() {
   const privateKey = privateDerBytes.slice(16, 16 + 32);
   return {publicKey, privateKey};
 }
+
+let deriveSecret, generateKeyPair;
+if(crypto.diffieHellman) {
+  deriveSecret = deriveSecretNative;
+  generateKeyPair = generateKeyPairNative;
+} else {
+  deriveSecret = cryptoNacl.deriveSecret;
+  generateKeyPair = cryptoNacl.generateKeyPair;
+}
+
+export {deriveSecret, generateKeyPair};
